added sorting, fixed nav_bar, converted weight

+++ lib/services/cloud/cloud_exercise_history.dart
@@ -2,7 +2,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 
 class CloudExerciseHistory {
   final String documentId;
-  final int weight;
+  final double weight;
   final int reps;
   final DateTime timestamp;
 
@@ -19,7 +19,7 @@ class CloudExerciseHistory {
     final data = snapshot.data()!;
     return CloudExerciseHistory(
       documentId: snapshot.id,
-      weight: data['weight'] as int,
+      weight: data['weight'] as double,
       reps: data['reps'] as int,
       timestamp: (data['timestamp'] as Timestamp?)?.toDate() ?? DateTime.now(),
     );
